requirement.txt
Django==2.0
filetype==1.0.5
gitdb2==2.0.5
GitPython==2.1.11
pytz==2018.9
smmap2==2.0.5
